Prefer the channel up button to the power button.
On AN-MR19BA, the power button doesn't seem to generate any BLE events but it does shoot out IR.
In a pinch, KEY_SLEEP, to the right of the power button can be repurposed to act as a power button.

Also add Makefile directive to allow compilation of the floating point code for Raspberry Pi.